Add files via upload
In this project, we deal with datasets of Movie consists of rating.dat, movie.dat and users.dat files. Spark RDD, Spark-SQL API, and MLLIB library are used to execute data frames queries and SQL queries on these files. In this mini-project we can count the max, min ratings along with the number of users who have rated a movie. Display the title, max rating, min rating, number of users, get movie titles to show with recommendations. To get predicted ratings to compare to test ratings, etc. This is the recommendation system that helps to strengthen the basics of mllib library of apache spark.